// src/game/data/enums.ts
export const enum ECombatEffectActorType {
  NONE = 0,
  SELF = 1,
  TARGET = 2,
  CASTER = 3
}

export const enum ECombatEffectConditionType {
  NONE = 0,
  CURRENT_SKILL = 1,
  HP_LESS = 2,
  HP_GREATER = 3,
  MP_LESS = 4,
  MP_GREATER = 5,
  NPC_GRADE_LESS = 6,
  NPC_GRADE_GREATER = 7,
  NPC_GRADE_EQUAL = 8,
  NPC_SPECIES = 9,
  NPC_ATTR = 10,
  ABNORMAL_MOVE = 11,
  ABNORMAL_STATUS = 12,
  ABNORMAL_MOVE_IMMUNE = 13,
  ABNORMAL_STATUS_IMMUNE = 14,
  ABNORMAL_MOVE_ALL = 15,
  PC = 16,
  SKILL_EFFECT_ID = 17,
  IDENTITY_STACK_COUNT = 18,
  STATUS_EFFECT_IMMUNETYPE = 19,
  ABNORMAL_NOT_MOVE = 20,
  TARGET_COUNT = 21,
  SKILL_IDENTITY_CATEGORY = 22,
  IDENTITY_ELEMENT_VALUE = 23,
  DIRECTIONAL_ATTACK = 24,
  CURRENT_SKILL_GROUP = 25,
  ABNORMAL_MOVE_STATUS_ALL = 26,
  IDENTITY_STANCE = 27,
  PC_SKILL = 28,
  SKILL_EFFECT_GROUP_SET = 29,
  NPC_ID = 30,
  IDENTITY_GAUGE0_VALUE_LESS = 31,
  PC_WITHOUT_ME = 32,
  NPC_SCALED_LEVEL_EQUAL = 33,
  NPC_SCALED_LEVEL_LESS = 34,
  NPC_SCALED_LEVEL_GREATER = 35,
  NOT_SKILL_EFFECT_ID = 36,
  ABNORMAL_MOVE_NOT_IMMUNE = 37,
  APPLY_TARGET_MARKING = 38,
  DAMAGE_ATTR = 39,
  IDENTITY_ELEMENT_VALUE_LESS = 40,
  COMMAND_SKILL_TYPE = 41,
  STATUS_EFFECT_ON = 42,
  IDENTITY_GAUGE1_VALUE_LESS = 43,
  HIT_RANGE_OVER = 44,
  HIT_RANGE_UNDER = 45,
  DIRECTIONAL_SKILL_EFFECT = 46,
  IDENTITY_GAUGE0_VALUE_GREATER = 47,
  ULTIMATE_SKILL_POINT_GREATER = 48,
  ULTIMATE_SKILL_POINT_LESS = 49
}

export const enum ECombatEffectActionType {
  NONE = 0,
  MODIFY_DAMAGE = 1,
  MODIFY_FINAL_DAMAGE = 2,
  MODIFY_CRITICAL_RATIO = 3,
  MODIFY_CRITICAL_MULTIPLIER = 4,
  MODIFY_PENETRATION = 5,
  MODIFY_PENETRATION_WHEN_CRITICAL = 6,
  EXEC_ACTIVE_EFFECT_WHEN_DAMAGE = 7,
  EXEC_ACTIVE_EFFECT_WHEN_CRITICAL = 8,
  EXEC_REACTIVE_EFFECT_WHEN_MISS = 9,
  EXEC_REACTIVE_EFFECT_WHEN_DAMAGE = 10,
  EXEC_REACTIVE_EFFECT_WHEN_CRITICAL = 11,
  EXEC_AFTER_EFFECT = 12,
  EXEC_AFTER_SKILL = 13,
  APPLY_HEAL = 14,
  MODIFY_REACTIVE_DAMAGE = 15,
  MODIFY_DAMAGE_SHIELD_MULTIPLIER = 16,
  EXEC_ACTIVE_EFFECT_WHEN_KILL = 17,
  EXEC_START_SKILL = 18,
  MODIFY_PENETRATION_ADDEND = 19,
  MODIFY_PENETRATION_ADDEND_WHEN_CRITICAL = 20,
  MODIFY_REACTIVE_CRITICAL_MULTIPLIER = 21,
  MODIFY_DAMAGE_WHEN_CRITICAL = 22,
  MODIFY_PARALYZATION_POINT = 23,
  EXEC_WHEN_COUNTER = 24,
  EXEC_WHEN_BE_KILLED = 25,
  MODIFY_EVOLUTION_DAMAGE_MULTIPLIER = 26,
  MODIFY_AWAKENING_DAMAGE = 27
}
// ...
export enum EInputTimingType {
  ITT_NONE = 0,
  ITT_MOVECANCEL = 1,
  ITT_INPUTCANCEL = 2,
  ITT_STAGECHANGE = 3,
  ITT_TIMINGCANCEL = 4,
  ITT_OVERLAPSKILL = 5,
  ITT_PERFECTZONE = 6,
  ITT_MOVEINPUT = 7,
  ITT_AUTOMOVEINPUT = 8,
  ITT_MOVESPEED = 9,
  ITT_GETUP_SKILL = 10,
  ITT_AUTO_SKILL_DIR_INPUT = 11,
  ITT_MAX = 12
}